requirements.txt
pytest==8.4.2
selenium==4.23.1
allure-pytest==2.15.3
requests==2.32.3
pytest-xdist==3.3.0
pytest-html==4.0.0
pytest-metadata==3.1.1
